Changed unsigned ints to unsigned longs where pointer arithmetic is involved. This eases life on our opteron cluster and shouldnt hurt the normal 32 bit pentiums

+++ include/qdp_outer.h
@@ -1,5 +1,5 @@
 // -*- C++ -*-
-// $Id: qdp_outer.h,v 1.38 2005-02-28 16:46:37 bjoo Exp $
+// $Id: qdp_outer.h,v 1.39 2005-06-15 14:33:18 bjoo Exp $
 
 #include "qdp_config.h"
 
@@ -412,7 +412,7 @@ private:
 	QDP_error_exit("Unable to new memory in alloc mem in qdp_outer.h");
       }
 #endif
-      F = (T*)(((unsigned int)F_orig + (QDP_ALIGNMENT_SIZE-1)) & ~(QDP_ALIGNMENT_SIZE-1));
+      F = (T*)(((unsigned long)F_orig + (QDP_ALIGNMENT_SIZE-1)) & ~(QDP_ALIGNMENT_SIZE-1));
 
 #if QDP_DEBUG >= 1
       QDP_info("%s OLattice_orig=0x%x, OLattice[%d]=0x%x, this=0x%x, bytes/site=%d",
